Separate model file into separate files
