+++ OptusWeather/OptusWeather/City.swift
@@ -8,70 +8,6 @@
 
 import Foundation
 
-/// Location
-public struct Location {
-    /// latitude
-    let latitude:Double
-    /// longitude
-    let longitude:Double
-}
-
-// Temperature
-public struct Temperature {
-    
-    /// Temperature in Celcius
-    let value:Float
-    /// Min temperature in Celcius
-    let minimum:Float
-    /// Max temperature in Celcius
-    let maximum:Float
-}
-
-/// Condition summary
-public struct Condition {
-    /// Name (Rain, Snow, Extreme etc.)
-    let name:String
-    /// Description
-    let description:String
-    /// Icon URL
-    let iconUrl:String
-}
-
-/// Wind details
-public struct Wind {
-    /// Wind speed in m/s
-    let speed:Float
-    /// Wind direction, degrees (meteorological)
-    let direction: Float
-}
-
-/// Weather info
-public struct Weather {
-    /// Temperature
-    let temperature:Temperature
-    /// Condition
-    let condition:Condition
-    /// Wind
-    let wind: Wind
-    /// Pressure in hPa
-    let pressure : Float
-    /// Humidity in %
-    let humidity : Float
-    /// % Cloudiness
-    let cloudiness: Float
-    /// Rain Volume (unknown units) in last 3h
-    let rainVolume: Float
-    /// Visibility in meters
-    let visibility:Int64
-}
-
-public struct Sun {
-    /// Time of Sunrise in local time
-    let sunrise: Date
-    /// Time of Sunset in local time
-    let sunset: Date
-}
-
 /// City with weather, location and sun details
 public struct City {
     /// City ID
