+++ OptusWeather/OptusWeather/Condition.swift
@@ -1,0 +1,19 @@
+//
+//  Condition.swift
+//  OptusWeather
+//
+//  Created by Vijay Santhanam on 24/05/2017.
+//  Copyright © 2017 Vijay Santhanam. All rights reserved.
+//
+
+import Foundation
+
+/// Condition summary
+public struct Condition {
+    /// Name (Rain, Snow, Extreme etc.)
+    let name:String
+    /// Description
+    let description:String
+    /// Icon URL
+    let iconUrl:String
+}
