+++ OptusWeather/OptusWeather/Location.swift
@@ -1,0 +1,17 @@
+//
+//  Location.swift
+//  OptusWeather
+//
+//  Created by Vijay Santhanam on 24/05/2017.
+//  Copyright © 2017 Vijay Santhanam. All rights reserved.
+//
+
+import Foundation
+
+/// Location
+public struct Location {
+    /// latitude
+    let latitude:Double
+    /// longitude
+    let longitude:Double
+}
