+++ OptusWeather/OptusWeather/Sun.swift
@@ -1,0 +1,17 @@
+//
+//  Sun.swift
+//  OptusWeather
+//
+//  Created by Vijay Santhanam on 24/05/2017.
+//  Copyright © 2017 Vijay Santhanam. All rights reserved.
+//
+
+import Foundation
+
+/// Details of sunrise and sunset
+public struct Sun {
+    /// Time of Sunrise in local time
+    let sunrise: Date
+    /// Time of Sunset in local time
+    let sunset: Date
+}
