+++ OptusWeather/OptusWeather/Temperature.swift
@@ -1,0 +1,20 @@
+//
+//  Temperature.swift
+//  OptusWeather
+//
+//  Created by Vijay Santhanam on 24/05/2017.
+//  Copyright © 2017 Vijay Santhanam. All rights reserved.
+//
+
+import Foundation
+
+// Temperature
+public struct Temperature {
+    
+    /// Temperature in Celcius
+    let value:Float
+    /// Min temperature in Celcius
+    let minimum:Float
+    /// Max temperature in Celcius
+    let maximum:Float
+}
