+++ OptusWeather/OptusWeather/Wind.swift
@@ -1,0 +1,17 @@
+//
+//  Wind.swift
+//  OptusWeather
+//
+//  Created by Vijay Santhanam on 24/05/2017.
+//  Copyright © 2017 Vijay Santhanam. All rights reserved.
+//
+
+import Foundation
+
+/// Wind details
+public struct Wind {
+    /// Wind speed in m/s
+    let speed:Float
+    /// Wind direction, degrees (meteorological)
+    let direction: Float
+}
